refactor(bundle): replace jspm api with systemjs-builder for both template and js
